- dispatch fix listeners with multiple slot indexes - made ID changes synchronously - added flushHiPriorityQueue to script thread queue - added unit tests for VoiceBitMap & multi slot listeners - fixed hardcoded midi processors not setting their attribute count

+++ hi_scripting/scripting/api/ScriptBroadcaster.h
@@ -484,7 +484,7 @@ private:
 	{
 		struct ProcessorListener;
 
-		ModuleParameterListener(ScriptBroadcaster* b, const Array<WeakReference<Processor>>& processors, const Array<int>& parameterIndexes, const var& metadata, const Identifier& specialId, bool useIntegerParameters);
+		ModuleParameterListener(ScriptBroadcaster* b, const Array<WeakReference<Processor>>& processors, const Array<uint16>& parameterIndexes, const var& metadata, const Identifier& specialId, bool useIntegerParameters);
 
 		Identifier getItemId() const override;
